chore: run prettier and fix imports for modular personal account feature
